backend/requirements.txt
asgiref==3.3.1
Django==3.1.4
pytz==2020.4
sqlparse==0.4.1

djangorestframework~=3.12.2
djangorestframework-jwt~=1.11.0